requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
cryptography==42.0.7
django-sendgrid-v5==1.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.2
sendgrid==6.11.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
